beginning to sort files into new structure
