+++ content/dungeon.js
@@ -1,4 +1,4 @@
-const secrets = require('./secrets-loader.js')
+const secrets = require('../secrets-loader.js')
 
 const strings = [
   ['Geistige Gesundheit', 'Sanity'],
